refactor: extract concurrency control and business policy from infra to zero
Move business policy functions (TIER_CONCURRENCY_LIMITS, getConcurrencyLimitForTier,
getEffectiveConcurrencyLimit, PENDING_RUN_TTL_MS, checkRunConcurrencyLimit,
authorizeCompose, validateComposeRequirements) from infra/run/run-service.ts to
a new zero-run-policy.ts in the zero layer.

This is a pure code relocation with no behavioral changes — the foundational
step that enables the subsequent transaction restructure.

Closes #8363
